refactor: move node status methods in a new file
